Fetch data from bajacalifornia
In a nutshell:
The page is too long and have too many html objects to be processed in
a reasonable time.
Fetches web page with fetch-clj because a string is required to trim the
HTML because it is too long. A map is inserted in the code as a
script.
Then the string is converted to a list of lines and lazy trimmed to
925 lines long.
The resulting list is converted back to a string, and then to a java
string object to be recognized by enlive/html-resource.
Then a enlive/select is applied.

-----
It works but I feel clumsy and dirty for all the handling.
Happy for using lazy sequences to reduce the number of lines.
No music. I was completely absorbed.
Now I am tired and intellectually excited.